finish refactor, about to begin testing
