+++ components/__tests__/InviteFriend.test.js
@@ -1,0 +1,3 @@
+import React from 'react';
+import renderer from 'react-test-renderer';
+
